requirements.txt
streamlit_option_menu
grad-cam
imageio
opencv-contrib-python-headless
numpy
pandas
streamlit-aggrid==0.2.3.post2
plotly
tensorflow==2.4.1
matplotlib
imutils
scikit-image
streamlit_webrtc
joblib
mtcnn
dlib-binary
